Move TabName to its own file to avoid having to mock it in my next commit
