+++ src/lib/Components/Artist/ArtistConsignButton.tsx
@@ -6,7 +6,8 @@ import { useTracking } from "react-tracking"
 import styled from "styled-components/native"
 
 import { ArtistConsignButton_artist } from "__generated__/ArtistConsignButton_artist.graphql"
-import { TabName, useSelectedTab } from "lib/NativeModules/SelectedTab/SelectedTab"
+import { useSelectedTab } from "lib/NativeModules/SelectedTab/SelectedTab"
+import { TabName } from "lib/NativeModules/SelectedTab/TabName"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { Router } from "lib/utils/router"
 import { Schema } from "lib/utils/track"
