+++ src/lib/NativeModules/SelectedTab/TabName.ts
@@ -1,0 +1,9 @@
+export enum TabName {
+  ARHomeTab = "ARHomeTab",
+  ARSearchTab = "ARSearchTab",
+  ARMessagingTab = "ARMessagingTab",
+  ARLocalDiscoveryTab = "ARLocalDiscoveryTab",
+  ARFavoritesTab = "ARFavoritesTab",
+  ARSalesTab = "ARSalesTab",
+  Unknown = "Unknown",
+}
